Add code for loads of stuff
The UZModelHolder can be set to be pushable, with some new enemies, items, and such other things.

+++ Sources/Engine/Base/Memory.cpp
@@ -114,7 +114,7 @@ void *AllocMemoryAligned(SLONG memsize, SLONG slAlignPow2)
 }
 void FreeMemoryAligned( void *memory)
 {
-    FreeMemory((void*)(((uintptr_t*)memory)[-1]));
+    FreeMemory((void*)(((ULONG*)memory)[-1]));
 }
 
 void FreeMemory( void *memory )
